Refactored Zone DB POST createZone endpoint.
